+++ groups/group_delete.go
@@ -1,0 +1,26 @@
+package groups
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/fivetran/go-fivetran/common"
+	httputils "github.com/fivetran/go-fivetran/http_utils"
+)
+
+type GroupDeleteService struct {
+	httputils.HttpService
+	groupID *string
+}
+
+func (s *GroupDeleteService) GroupID(value string) *GroupDeleteService {
+	s.groupID = &value
+	return s
+}
+
+func (s *GroupDeleteService) Do(ctx context.Context) (common.CommonResponse, error) {
+	var response common.CommonResponse
+	url := fmt.Sprintf("/groups/%v", *s.groupID)
+	err := s.HttpService.Do(ctx, "DELETE", url, nil, nil, 200, &response)
+	return response, err
+}